# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "AcquiredFulltext": {
        "description": "已下载、校验并写入私有暂存区的 PDF 信息。",
        "properties": {
          "access_rights": {
            "default": "open_access",
            "title": "Access Rights",
            "type": "string"
          },
          "acquired_at": {
            "format": "date-time",
            "title": "Acquired At",
            "type": "string"
          },
          "byte_size": {
            "exclusiveMinimum": 0.0,
            "title": "Byte Size",
            "type": "integer"
          },
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "doi": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Doi",
            "type": "string"
          },
          "media_type": {
            "default": "application/pdf",
            "title": "Media Type",
            "type": "string"
          },
          "origin_kind": {
            "default": "open_access",
            "title": "Origin Kind",
            "type": "string"
          },
          "original_filename": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Original Filename",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "source_url": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          },
          "staging_object_key": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Staging Object Key",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "doi",
          "source_url",
          "staging_object_key",
          "original_filename",
          "byte_size",
          "sha256",
          "acquired_at"
        ],
        "title": "AcquiredFulltext",
        "type": "object"
      },
...
      "CandidateFulltextResponse": {
        "description": "单篇候选全文任务的短期状态；文件详情仅在 available 时出现。",
        "properties": {
          "attempt_no": {
            "title": "Attempt No",
            "type": "integer"
          },
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AcquiredFulltext"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FulltextAcquisitionError"
              },
              {
                "type": "null"
              }
            ]
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "search_run_id": {
            "format": "uuid",
            "title": "Search Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FulltextAcquisitionStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "search_run_id",
          "candidate_id",
          "attempt_no",
          "status",
          "requested_at",
          "updated_at"
        ],
        "title": "CandidateFulltextResponse",
        "type": "object"
      },
...
      "CollectionAdmissionResult": {
        "description": "加入成功或幂等命中后，供 API 与后续 Worker 使用的标识集合。",
        "properties": {
          "collection_id": {
            "format": "uuid",
            "title": "Collection Id",
            "type": "string"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "ingestion_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingestion Run Id"
          },
          "paper_id": {
            "format": "uuid",
            "title": "Paper Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CollectionAdmissionStatus"
          }
        },
        "required": ["status", "collection_id", "paper_id"],
        "title": "CollectionAdmissionResult",
        "type": "object"
      },
      "CollectionAdmissionStatus": {
        "description": "文献加入集合后的幂等结果。",
        "enum": ["added", "already_joined"],
        "title": "CollectionAdmissionStatus",
        "type": "string"
      },
...
      "CreateWorkspaceRequest": {
        "description": "新建研究工作区的最小输入，不允许客户端指定所有者或状态。",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": ["name"],
        "title": "CreateWorkspaceRequest",
        "type": "object"
      },
...
      "FulltextAcquisitionError": {
        "description": "不包含响应正文、Cookie 或临时签名 URL 的安全失败摘要。",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/FulltextAcquisitionErrorCode"
          },
          "http_status_code": {
            "anyOf": [
              {
                "maximum": 599.0,
                "minimum": 100.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Http Status Code"
          },
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": ["code", "message", "retryable"],
        "title": "FulltextAcquisitionError",
        "type": "object"
      },
      "FulltextAcquisitionErrorCode": {
        "description": "全文获取失败的可展示、可统计原因码。",
        "enum": [
          "citation_not_ready",
          "missing_doi",
          "doi_mismatch",
          "not_open_access",
          "missing_fulltext_url",
          "invalid_url",
          "unsafe_url",
          "redirect_limit_exceeded",
          "timeout",
          "network_error",
          "remote_error",
          "file_too_large",
          "invalid_content_type",
          "invalid_pdf",
          "storage_error",
          "task_error"
        ],
        "title": "FulltextAcquisitionErrorCode",
        "type": "string"
      },
      "FulltextAcquisitionStatus": {
        "description": "全文获取任务对调用方暴露的可恢复状态。",
        "enum": [
          "queued",
          "downloading",
          "validating",
          "available",
          "requires_upload",
          "rejected",
          "failed"
        ],
        "title": "FulltextAcquisitionStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowStageDisplay": {
        "description": "工作区研究阶段的中文展示元数据，不作为客户端提交输入。",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": ["label", "description"],
        "title": "WorkflowStageDisplay",
        "type": "object"
      },
      "WorkspaceListResponse": {
        "description": "工作区切换器使用的游标分页结果。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": ["items"],
        "title": "WorkspaceListResponse",
        "type": "object"
      },
      "WorkspaceResponse": {
        "description": "工作区列表和详情共用的可展示字段。",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "research_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Question"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow_stage": {
            "$ref": "#/components/schemas/WorkspaceWorkflowStage"
          },
          "workflow_stage_display": {
            "$ref": "#/components/schemas/WorkflowStageDisplay"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "research_question",
          "status",
          "workflow_stage",
          "workflow_stage_display",
          "created_at",
          "updated_at"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      },
      "WorkspaceWorkflowStage": {
        "description": "工作区在研究主流程中的阶段；英文值会写入数据库和 API。",
        "enum": [
          "draft",
          "analyzing",
          "plan_review",
          "retrieving",
          "screening",
          "collection_building",
          "researching",
          "failed"
        ],
        "title": "WorkspaceWorkflowStage",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "面向学术文献检索与 RAG 研究工作区的后端服务。",
    "title": "Academic Search API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/collections": {
      "get": {
        "description": "供工作区切换器按需加载，默认只显示活动工作区。",
        "operationId": "list_workspaces_api_v1_collections_get",
        "parameters": [
          {
            "description": "是否包含已归档工作区",
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "description": "是否包含已归档工作区",
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "description": "按工作区名称或当前研究阶段搜索",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按工作区名称或当前研究阶段搜索",
              "title": "Q"
            }
          },
          {
            "description": "上一页返回的不透明分页游标",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "上一页返回的不透明分页游标",
              "title": "Cursor"
            }
          },
          {
            "description": "本次最多返回的工作区数量",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "本次最多返回的工作区数量",
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "搜索我的研究工作区",
        "tags": ["研究工作区"]
      },
      "post": {
        "description": "创建一个空工作区；研究问题将在进入工作区后另行确认。",
        "operationId": "create_workspace_api_v1_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "创建研究工作区",
        "tags": ["研究工作区"]
      }
    },
...
    "/api/v1/collections/{collection_id}/search-runs/{search_run_id}/candidates/{candidate_id}/fulltext": {
      "get": {
        "description": "供前端轮询正在下载、校验或已可加入集合的候选全文。",
        "operationId": "get_candidate_fulltext_api_v1_collections__collection_id__search_runs__search_run_id__candidates__candidate_id__fulltext_get",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "search_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Search Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateFulltextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "获取候选全文任务状态",
        "tags": ["全文获取"]
      },
      "post": {
        "description": "创建全文异步任务；候选、URL 和题录均从服务端检索会话读取。",
        "operationId": "request_candidate_fulltext_api_v1_collections__collection_id__search_runs__search_run_id__candidates__candidate_id__fulltext_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "search_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Search Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateFulltextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "获取候选的开放获取全文",
        "tags": ["全文获取"]
      }
    },
    "/api/v1/collections/{collection_id}/search-runs/{search_run_id}/candidates/{candidate_id}/fulltext/admission": {
      "post": {
        "description": "只接受当前搜索会话中已校验 PDF，创建不会被 Worker 提前领取的 pending 运行。",
        "operationId": "admit_candidate_fulltext_api_v1_collections__collection_id__search_runs__search_run_id__candidates__candidate_id__fulltext_admission_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "search_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Search Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionAdmissionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "将已验证全文加入待确认集合",
        "tags": ["全文获取"]
      }
    },
    "/api/v1/collections/{collection_id}/search-runs/{search_run_id}/candidates/{candidate_id}/fulltext/retry": {
      "post": {
        "description": "仅为可重试的终态失败创建新的全文下载尝试。",
        "operationId": "retry_candidate_fulltext_api_v1_collections__collection_id__search_runs__search_run_id__candidates__candidate_id__fulltext_retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "search_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Search Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateFulltextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "重试候选全文获取",
        "tags": ["全文获取"]
      }
    },
    "/api/v1/collections/{collection_id}/search-runs/{search_run_id}/candidates/{candidate_id}/fulltext/upload": {
      "post": {
        "description": "只接收二进制 PDF 流；授权、候选、DOI 和暂存对象键都由服务端控制。",
        "operationId": "upload_candidate_fulltext_api_v1_collections__collection_id__search_runs__search_run_id__candidates__candidate_id__fulltext_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Collection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "search_run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Search Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-upload-authorized",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Upload-Authorized"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateFulltextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "上传有权处理的候选 PDF",
        "tags": ["全文获取"]
      }
    },
...
  }
}